refactor for the event listener done, but its broken, got to fix it but the structure should be what we want
