+++ static/js/tournaments/populateHelpers.js
@@ -1,0 +1,185 @@
+
+//Build TREE
+const buildTr = (matchInfos) => {
+    let left, right = ``
+
+    if (matchInfos.status == "creating"){
+        left = `
+                <img src="/static/img/huh.webp" alt="">
+                <div>-- tbd --</div>
+        `
+        right = `
+                <img src="/static/img/huh.webp" alt="">
+                <div>-- tbd --</div>
+        `
+
+    }else if (matchInfos.status == "finished"){
+
+        left = `
+                <img src="${matchInfos.winner_picture}" alt="">
+                <div>${matchInfos.winner_username}</div>
+        `
+
+        right = `
+                <img src="${matchInfos.loser_picture}" alt="">
+                <div>${matchInfos.loser_username}</div>
+        `
+    }else {
+        left = `
+            ${`<img src="${matchInfos.playerL_picture}" alt="">`}
+            <div>${matchInfos.playerL_username}</div>
+        `
+
+        right = `
+            ${`<img src="${matchInfos.playerL_picture}" alt="">`}
+            <div>${matchInfos.playerL_username}</div>
+        `
+    }
+    return `<tr>
+                <td class="players">
+                    <div class="left">
+                        ${left}
+                    </div>
+                    <div class="vs">VS</div>
+                    <div class="right">
+                        ${right}
+                    </div>
+                </td>
+                <td><span class="badge bg-success">${matchInfos.status}</span></td>
+                <td><span class="left-score winner">${matchInfos.winner_score}</span> - <span class="right-score loser">${matchInfos.loser_score}</span></td>
+            </tr>`
+}
+
+const buildRound = (roundInfos, name)=> {
+
+    console.log(roundInfos)
+    return `
+    <div class="round mt-5">
+        <h2>${name}</h2>
+        <table class="table mt-2">
+            <thead>
+            <tr>
+                <th scope="col">Players</th>
+                <th scope="col">Status</th>
+                <th scope="col">Scores</th>
+            </tr>
+            </thead>
+            <tbody>
+
+    ${Object.entries(roundInfos.games)
+        .map(([key, round]) => buildTr(round, key))
+        .join("")}
+            </tbody>
+        </table>
+    </div>
+    `
+}
+
+let meow = null
+
+export function populateTree(tournamentInfos) {
+        meow = document.querySelector("#tree")
+        if (meow == null) return ;
+        meow.innerHTML = '';
+        for (const key in tournamentInfos?.scoreboard.rounds) {
+            meow.innerHTML += buildRound(tournamentInfos?.scoreboard.rounds[key], key)
+        }
+}
+
+
+//TOURNAMENT ROOM
+export function populateTournament(tournament_data){
+    const max_clients = tournament_data.max_clients
+    const clientsInTournament = document.querySelector("#clientsInTournament");
+    let clientsDiv = []
+    clientsInTournament.innerHTML = ``
+    for (let i = 0 ; i < max_clients ; i++){
+        clientsDiv.push(`
+            <div class="col p-2">
+            <div class="content border p-3 d-flex justify-content-between align-items-center">
+            <div class="d-flex align-items-center justify-content-center gap-3">
+            <div class="nickname ml-3"> waiting</div>
+            <div>(add bpoyer's animation with the li dots)</div>
+            </div>
+            </div>
+            </div>
+            </div>
+            `)
+        }
+        
+        tournament_data?.players_infos?.forEach((player , i)=> {
+            clientsDiv[i] = `
+            <div class="col p-2">
+            <div class="content border p-3 d-flex justify-content-between align-items-center">
+                            <div class="d-flex align-items-center justify-content-center gap-3">
+                            <div class="avatarContainer">
+                                    <img src="${player.avatar}" alt="">
+                                    </div>
+                                <div class="nickname ml-3">${ player.nickname }</div>
+                            </div>
+
+                            <div class="d-flex justify-content-center align-items-center gap-1 flex-column">
+                            <img class="tropheeImg" src="${player.trophee}" alt="">
+                            <div>mmr: ${ player.mmr }</div>
+                            </div>
+
+                        </div>
+                        </div>
+                        `
+                        
+                        console.log("mini meow: ", clientsDiv[i])
+            });
+    clientsDiv.forEach(htmlString => {
+        const temp = document.createElement("div");
+        temp.innerHTML = htmlString; 
+
+        while (temp.firstChild) {
+            console.log(temp.firstChild)
+            clientsInTournament.appendChild(temp.firstChild);
+        }
+    });
+}
+
+///JOIN TOURNAMENT
+export function populateJoinTournament(tournamentList){
+    const tournaments = document.querySelector("#tournaments");
+    let clientsDiv = []
+    tournaments.innerHTML = ``
+    
+    tournamentList.forEach((room)=>{
+        clientsDiv.push(`
+        <div class="col room d-flex justify-content-center align-items-center" data-code="${room.code}" >
+            <div class="cardContainer">
+                <h4>${room.title}</h4>
+
+
+                <div class="d-flex flex-row align-items-center justify-content-between">
+                    <div class="btn joinTournamentBtn">join</div>
+                    <div class="avatarsContainer mt-2">
+                        ${room.players_infos.map((player, i)=>
+                            `
+                            <div class="imgContainer">
+                                <img src="${player.avatar}" alt="">
+                                ${(i == 2) ? `<div class="more">+${room.max_clients - (i + 1)}</div>`: ''}
+                            </div>
+                            `
+
+                        ).join('')}
+
+
+                    </div>
+                </div>
+            </div>
+        </div> 
+        `)
+    })
+    clientsDiv.forEach(htmlString => {
+        const temp = document.createElement("div");
+        temp.innerHTML = htmlString; 
+
+        while (temp.firstChild) {
+            console.log(temp.firstChild)
+            tournaments.appendChild(temp.firstChild);
+        }
+    });
+}
