+++ templates/base.html
@@ -375,6 +375,7 @@
     </main>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-MrcW6ZMFYlzcLA8Nl+NtUVF0sA7MsXsP1UyJoMp4YLEuNSfAP+JcXn/tWtIaxVXM" crossorigin="anonymous"></script>
     <script type="module" src="{% static 'js/spa/spa.js' %}" defer></script>
+
     <!-- <script>window.chatSocket = new WebSocket('ws://' + window.location.host + '/ws/chat/?id=' + clientId);</script> -->
   </body>
   </html>
